Merge 166 commits.

+++ R/compile_lesson.R
@@ -5,7 +5,7 @@
 #' Combines functionality of:
 #' - [compile_fm()]
 #' - [compile_standards()]
-#' - [learningChart()] and [learningEpaulette()]
+#' - [learningEpaulette()]
 #' - [compileAcknowledgments()]
 #' - [compileVersions()]
 #' - [compileJSON()]
@@ -14,8 +14,6 @@
 #' @param WD is working directory of the project (useful to supply for shiny app, which has diff. working environment); if "?" supplied, will invoke [pick_lesson()]
 #' @param choices one or more of the following: c("Front Matter","Standards Alignment","Teaching Materials","Procedure","Acknowledgements","Versions"); or "All". If missing, will compile things in the ReadyToCompile entry in front-matter.yml for the WD folder.
 #' @param current_data the reconciled data including yaml and input from the shiny app environment; if current_data=NULL, read in front-matter.yml
-#' @param destFolder where you want to save the folder; by default in the "meta/JSON/" folder
-#' @param outputFileName output file name; default= "processedProcedure.json"
 #' @param clean delete all JSON files in meta/ and start over? default=FALSE
 #' @param rebuild if T, rebuild everything; overrides RebuildAllMaterials in front-matter.yml; default= NULL
 #' @return current_data; also the lesson JSON is saved to `meta/JSON/LESSON.json`
@@ -23,21 +21,19 @@
 #' @export
 #'
 compile_lesson <-
-  function(WD = getwd(),
+  function(WD = "?",
            choices,
            current_data,
-           destFolder ,
-           outputFileName = "LESSON.json",
            clean = FALSE,
            rebuild = NULL) {
     WD <- parse_wd(WD)
 
-    #figure out which repo we're connected to (to create full paths to catalog.galacticpolymath.com)
-    repo <- whichRepo(WD = WD)
 
 
     # Always update front-matter (in case of template updates) ----------------
-    update_fm(WD = WD, save_output = TRUE)
+    update_fm(WD = WD, save_output = TRUE,recompile = FALSE)
+    #run upload_assets to make sure there's nothing new
+    upload_assets(WD = WD)
 
     if (missing(current_data)) {
       current_data <-
@@ -47,9 +43,16 @@ compile_lesson <-
     if (missing(choices)) {
       choices <- current_data$ReadyToCompile
     }
-    if (missing(destFolder)) {
-      destFolder <- fs::path(WD, "meta", "JSON")
-    }
+
+
+
+    # Find the path for the GitHub gp-lessons working dir ---------------------
+    WD_git <- get_wd_git(WD = WD)
+
+    destFolder <- fs::path(WD_git, "JSONs")
+
+    proj_dir <- get_fm("GdriveDirName", WD = WD)
+    checkmate::assert_character(proj_dir, min.chars = 2)
 
     if (!dir.exists(destFolder)) {
       stop("Directory not found: ", destFolder)
@@ -69,8 +72,12 @@ compile_lesson <-
     if (clean) {
       to_delete <-
         list.files(destFolder, pattern = "*\\.json", full.names = TRUE)
-      unlink(to_delete)
-      message("\nFolder cleared: ", destFolder, "\n")
+      if (length(to_delete) > 0) {
+        unlink(to_delete)
+        message("\nJSONs folder cleared: ", destFolder, "\n")
+      } else{
+        message("\nNo JSONs found at: ", destFolder, "\n")
+      }
     }
 
 
@@ -88,14 +95,15 @@ compile_lesson <-
     }
 
     #quell Rcheck
-    lumpItems <- whichRepo <- catalogURL <- expand_md_links <- NULL
+    lumpItems <- expand_md_links <- NULL
 
 
     # Define some paths -------------------------------------------------------
     status <- get_fm("PublicationStatus", WD = WD, checkWD = F)
     med_title <- get_fm("MediumTitle", WD = WD, checkWD = F)
 
-    checkmate::assert_choice(status, c("Live", "Draft"))
+    checkmate::assert_choice(status,
+                           c("Proto","Hidden","Beta","Coming Soon", "Live","Draft"))#draft deprecated
     checkmate::assert_character(med_title, min.chars = 2)
 
     # local path to teaching material
@@ -103,12 +111,14 @@ compile_lesson <-
     # Else, found on 'GalacticPolymath'
     #****To do::: need to construct this from shared_drive_path, GdriveTeachMatPath
 
-    tm_path_full <-fs::path(get_shared_drive_path(),get_fm("GdriveTeachMatPath",WD=WD,checkWD=F))
+    tm_path_full <-
+      fs::path(get_shared_drive_path(),
+               get_fm("GdriveTeachMatPath", WD = WD, checkWD = F))
 
     #make sure we know local directory path to this lesson's teaching-materials
     #In case we're waiting on Google Drive for desktop to update, let's repeat this assertion after it fails, increasing wait time
     checkmate::assert_directory_exists(tm_path_full, .var.name = "GdriveTeachMatPath") %>%
-      catch_err(try_harder = T, waits = c(2, 5, 10, 15))
+      catch_err(try_harder = T, waits = c(2, 5, 10, 15, 5))
 
 
 
@@ -116,7 +126,8 @@ compile_lesson <-
     # test if learningEpaulette is up-to-date with the 'standards_ShortTitle.gsheet' file, or if any of these files is missing.
 
     compiled_standards_path <-
-      fs::path(WD, "meta", "standards.RDS")
+      fs::path(WD_git, "saves", "standards.RDS")
+
 
     standards_gsheet_path <- fs::path(WD,
                                       "meta",
@@ -124,7 +135,7 @@ compile_lesson <-
                                       ext = "gsheet")
 
     compiled_standards_json_path <-
-      fs::path(WD, "meta", "json",  "standards.json")
+      fs::path(destFolder,  "standards.json")
 
     teach_it_path <- fs::path(WD,
                               "meta",
@@ -133,33 +144,32 @@ compile_lesson <-
 
     #compiled standards should be newer than standards gsheet
 
-    stnds_out_of_date <- !inSync(compiled_standards_path,
-                                 standards_gsheet_path,
-                                 newer = TRUE)
+    stnds_out_of_date <- !inSync(
+      compiled_standards_json_path,
+      compiled_standards_path,
+      standards_gsheet_path,
+      newer = TRUE,
+      WD = WD
+    )
 
     # Compile standards if out of date or missing or rebuild==T ----------------
     if ("Standards Alignment" %in% choices &
         (stnds_out_of_date | rebuild)) {
-      message("Recompiling standards to reflect newer 'teach_it*.gsheet'")
+      message("Recompiling standards to reflect newer 'standards*.gsheet'")
       compile_standards_output <- compile_standards(
         WD = WD,
-        targetSubj = current_data$TargetSubject,
-        learningplot_correction = current_data$LearningPlotCorrection
+        targetSubj = current_data$TargetSubject
       ) %>% catch_err(keep_results = T)
 
-      if(!compile_standards_output$success){
+      if (!compile_standards_output$success) {
         stop("Standards were not compiled successfully.")
+      } else{
+        alignment <- compile_standards_output$result
+
       }
-      alignment <- compile_standards_output$result
-      current_data$LearningChartFriendly <-
-        alignment$learning_chart_friendly
-      if (is.na(current_data$TargetSubject)) {
-        warning("Enter a Target Subject on the Edit tab and try again.")
-      }
-      message("\nGenerating Learning Chart\n")
     }
 
-
+    #??Is this why compile_lesson sometimes runs once unnecessarily?
     if ("Standards Alignment" %in% choices) {
       # Test if standards are compatible with learning chart --------------------
       if (!file.exists(compiled_standards_path)) {
@@ -175,92 +185,32 @@ compile_lesson <-
         save_json(sh,
                   fs::path(destFolder, "standards-header.json"))
 
-        save_json(
-          saved_standards$data$list_for_json,
-          fs::path(WD, "meta", "JSON", "standards.json")
-        )
-      }
-      #Only proceed to generate learningChart if compatible...
-      if (!saved_standards$learning_chart_friendly) {
-        current_data$LearningChart <- NULL
-      } else if (!inSync(
-        fs::path(WD, "assets", "_learning-plots", "GP-Learning-Chart.png"),
-        standards_gsheet_path
-      ) |
-      (rebuild)) {
-        #LEARNING CHART
-        learningChart(
-          quotedTitle = current_data$Title,
-          centralText = current_data$LearningChart_params_centralText,
-          caption = current_data$LearningChart_params_caption,
-          captionN = current_data$LearningChart_params_captionN,
-          showPlot = FALSE,
-          WD = WD
-        )
-
-        #set learning chart filename from default file output on learningChart function
-        #(since this file doesn't exist in yaml yet)
-        lcname <- fs::path("assets",
-                           "_learning-plots",
-                           paste0(formals(learningChart)$fileName, ".png"))
-        if (file.exists(lcname)) {
-          current_data$LearningChart <- lcname
-        }
+        save_json(saved_standards$data$list_for_json,
+                  fs::path(destFolder, "standards.json"))
       }
 
 
-      #If learning chart exists, always output the learning-chart.json
-      if (!is_empty(current_data$LearningChart) &
-          identical(TRUE, file.exists(fs::path(WD, current_data$LearningChart)))) {
-        #export learning chart section
-        lc <- list(
-          `__component` = "lesson-plan.learning-chart",
-          Title = "About the GP Learning Chart",
-          Description =
-            paste0(
-              "This Galactic Polymath Learning Chart illustrates the areas of knowledge covered. This lesson targets ",
-              current_data$TargetSubject,
-              ", but it helps teach national learning standards in 4 subjects: \n- [Common Core Math](http://www.corestandards.org/Math/); [Common Core ELA](http://www.corestandards.org/ELA-Literacy/); [Next Generation Science (NGSS)](https://www.nextgenscience.org/); and [College, Career, and Civic Life (C3) Social Studies Standards](https://www.socialstudies.org/standards/c3).\nIn total, there are ",
-              sum(saved_standards$a_combined$n, na.rm = T),
-              " standards across US grade band(s): ",
-              paste0(saved_standards$data$gradeBand, collapse = ', '),
-              "."
-            ),
-          Footnote = "",
-          Badge = list(url = ifelse(
-            is_empty(current_data$LearningChart[1]),
-            NA,
-            catalogURL(basename(current_data$LearningChart[1]), repo)
-          ))
 
-
-        )
-
-        #write learning chart section before standards section
-        jsonlite::write_json(
-          lc,
-          fs::path(destFolder, "learning-chart.json"),
-          pretty = TRUE,
-          auto_unbox = TRUE,
-          na = "null",
-          null = "null"
-        )
-      }
 
     }#end general standards stuff
+
+    ep_file <- fs::path(WD,
+                        "assets",
+                        "_learning-plots",
+                        "GP-Learning-Epaulette.png")
+    ep_vert_file <- fs::path(WD,
+                             "assets",
+                             "_learning-plots",
+                             "GP-Learning-Epaulette_vert.png")
 
     #Remake Epaulette if out of date or missing
     if ("Standards Alignment" %in% choices &
         (
-          !inSync(
-            fs::path(
-              WD,
-              "assets",
-              "_learning-plots",
-              "GP-Learning-Epaulette.png"
-            ),
-            standards_gsheet_path
-          ) | rebuild | is_empty(current_data$LearningEpaulette)
+          !inSync(ep_file,
+                  ep_vert_file,
+                  standards_gsheet_path,
+                  WD = WD) |
+          rebuild | is_empty(current_data$LearningEpaulette)
         )) {
       #####################
       #LEARNING EPAULETTE
@@ -273,16 +223,8 @@ compile_lesson <-
         randomSeed = current_data$LearningEpaulette_params_randomSeed
       ) %>% catch_err()
 
-      #set learning epaulette filename from default file output on learningEpaulette function
-      #(since this file doesn't exist in yaml on first run)
-      current_data$LearningEpaulette <-
-        fs::path("assets",
-                 "_learning-plots",
-                 paste0(formals(learningEpaulette)$fileName, ".png"))
-      current_data$LearningEpaulette_vert <-
-        fs::path("assets",
-                 "_learning-plots",
-                 paste0(formals(learningEpaulette)$fileName, "_vert.png"))
+
+
 
     }#end epaulette
 
@@ -293,49 +235,91 @@ compile_lesson <-
 
 
     if ("Teaching Materials" %in% choices) {
-      if (is.na(current_data$GitHubURL)) {
-        warning(
-          "GitHubURL is missing from front-matter.yml...if this doesn't work, try running update_fm()."
-        )
+      #check if previous save file exists
+      save_path <-
+        fs::path(WD_git, "saves", "save-state_teach-it.RDS")
+      test_save_exists <- file.exists(save_path)
+      if (!test_save_exists) {
+        skip_update <- FALSE
       } else{
-        #check if previous save file exists
-        save_path <- fs::path(WD, "meta", "save-state_teach-it.RDS")
-        save_exists <- file.exists(save_path)
-        if (save_exists) {
-          #compare current timestamps and file counts from last update to current
-          prev_update_state <- readRDS(save_path)
-          #get state for teach-it.gsheet AND all teaching-materials/ contents
-          curr_update_state <-
-            get_state(c(teach_it_path, tm_path_full), save_path = NULL)
+        #compare current timestamps and file counts from last update to current
+        prev_update_state <- readRDS(save_path)
+        #get state for teach-it.gsheet AND all teaching-materials/ contents
+        curr_update_state <-
+          get_state(c(teach_it_path, tm_path_full), save_path = NULL)
 
-          skip_update <-
-            identical(prev_update_state, curr_update_state)
+        if (identical(prev_update_state, curr_update_state)) {
+          skip_update <- TRUE
         } else{
           skip_update <- FALSE
         }
-
-        if (!skip_update | rebuild) {
-          # update teach_it links and compile ---------------------------------------
-          message("Changes to `../teaching-materials/` detected...")
-          message("Running update_teach_links() and compile_teach-it()")
-
-          test_update_teach_it <- update_teach_links(WD = WD) %>% catch_err()
-          test_compile_teach_it <- compile_teach_it(WD = WD) %>% catch_err()
-          #update the cache of the teaching-material state of things
-          get_state(path = c(teach_it_path, tm_path_full),
-                    save_path = save_path)
-        } else{
-          message("No changes to `../teaching-materials/` detected...")
-          message("Skipping update_teach_links() and compile_teach-it()")
-        }
       }
+      #can't just test "identical b/c there are some slight mod time diffs that are a pain in the ass
+      #to deal with (i.e. google drive might take a minute to sync)
+      #all old files exist in new folder
+      #   old_still_found <-
+      #     prev_update_state$path %in% curr_update_state$path %>% sum() == length(prev_update_state$path)
+      #   if (!old_still_found) {
+      #     skip_update <- FALSE
+      #   } else{
+      #     testL <- lapply(1:nrow(curr_update_state), \(i) {
+      #       curr_file_i <- curr_update_state[i, ]
+      #       matching_old_file_i <-
+      #         prev_update_state %>% dplyr::filter(.data$path == curr_file_i$path)
+      #
+      #       test <- nrow(matching_old_file_i) == 1
+      #       if (test) {
+      #         test2 <- test &
+      #           matching_old_file_i$size == curr_file_i$size &
+      #           #test whether sync time is within 3 minutes;
+      #           #helps deal with annoying difference in mod time b/w cloud and Gdrive for desktop
+      #           difftime(
+      #             curr_file_i$modification_time,
+      #             matching_old_file_i$modification_time ,
+      #             units = "mins"
+      #           ) <= 3
+      #       } else{
+      #         test2 <- test
+      #       }
+      #       test2
+      #     }) %>% unlist()
+      #     skip_update <- sum(testL) == length(testL)
+      #   }
+      # }else{
+      #   skip_update <- FALSE
+      # }
+
+      if (!skip_update | rebuild) {
+        # update teach_it links and compile ---------------------------------------
+        message("Changes to `../teaching-materials/` detected...")
+        message("Running update_teach_links() and compile_teach-it()")
+
+        test_update_teach_it <-
+          update_teach_links(WD = WD) %>% catch_err()
+        test_compile_teach_it <-
+          compile_teach_it(WD = WD) %>% catch_err()
+
+
+        #update the cache of the teaching-material state of things
+        get_state(
+          path = c(teach_it_path, tm_path_full),
+          save_path = save_path,
+          path1_modTime_diff = 2
+        )
+      } else{
+        message("No changes to `../teaching-materials/` detected...")
+        message("Skipping update_teach_links() and compile_teach-it()")
+      }
+
 
     }
 
 
     # Separate parts of Front Matter ------------------------------------------
+
     #always rebuild front matter if it's in choices
     if ("Front Matter" %in% choices) {
+      message("• Running compile_fm()")
       compile_fm(WD = WD)
 
     }#End of Front Matter export
@@ -343,24 +327,35 @@ compile_lesson <-
     # Printable Lesson --------------------------------------------------------
 
     if ("Printable Lesson" %in% choices) {
+      message("• Running make_printable()")
       make_printable(WD = WD, rebuild = rebuild)
 
     }
 
+
+# Make Shareable Assets ---------------------------------------------------
+message("• Running make_shareable_assets()")
+make_shareable_assets(WD=WD,open_file = FALSE)
+
+
+
     ################################################################
     # Compile all JSONs ----------------------------------------------
-    compileJSON(WD = WD)
+    message("• Running compile_json()")
+    compile_json(WD_git = WD_git)
 
-    #after run, reset rebuild all trigger
+    #after run, reset rebuild-all trigger
     if (rebuild) {
-      current_data$RebuildAllMaterials <- FALSE
+      message("• Running update_fm()")
+      update_fm(WD_git=WD_git,change_this = list(RebuildAllMaterials=FALSE))
+
     }
 
-    #Save updated YAML
-    yaml::write_yaml(current_data, fs::path(WD, "meta", "front-matter.yml"))
 
     invisible(current_data)
   }
+
+#alias
 
 #' lesson_compile
 #'
